+++ src/ui.h
@@ -14,6 +14,7 @@ class UI {
         int drawButton(Surface *s, const std::string &btn, const std::string &text, int x=5, int y=-10);
         int drawButtonRight(Surface *s, const std::string &btn, const std::string &text, int x=5, int y=-10);
         void drawScrollBar(uint32_t pagesize, uint32_t totalsize, uint32_t pagepos, SDL_Rect scrollRect);
+        static int intTransition(int from, int to, int32_t tickStart, int32_t duration = 500, int32_t tickNow = -1);
 };
 
 #endif